requirements.txt
streamlit
pandas
matplotlib
wordcloud
urlextract
emoji
vaderSentiment
Pillow 